requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
phonenumbers==8.13.18
psycopg2-binary==2.9.7
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
PyYAML==6.0.1
